+++ test/bacnet/basic/object/lc/src/main.c
@@ -34,7 +34,7 @@
  */
 static bool is_float_equal(float x1, float x2)
 {
-    return fabs(x1 - x2) < 0.001;
+    return fabsf(x1 - x2) < 0.001f;
 }
 
 /**
